+++ init/Storyboards/CategoryList.storyboard
@@ -28,7 +28,7 @@
         <!--Mission Category Table View Controller-->
         <scene sceneID="6jh-cf-h4B">
             <objects>
-                <tableViewController id="qL4-tR-ofK" customClass="MissionCategoryTableViewController" customModule="init" customModuleProvider="target" sceneMemberID="viewController">
+                <tableViewController id="qL4-tR-ofK" customClass="CategoryListController" customModule="init" customModuleProvider="target" sceneMemberID="viewController">
                     <tableView key="view" clipsSubviews="YES" contentMode="scaleToFill" alwaysBounceVertical="YES" dataMode="prototypes" style="plain" separatorStyle="default" rowHeight="44" sectionHeaderHeight="28" sectionFooterHeight="28" id="qzz-Xr-X6e">
                         <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
