+++ init/Storyboards/MissionList.storyboard
@@ -13,7 +13,7 @@
         <!--CPS Lab-->
         <scene sceneID="JM7-kd-Mcl">
             <objects>
-                <tableViewController id="1Z0-78-Skw" customClass="MissionListTableViewController" customModule="init" customModuleProvider="target" sceneMemberID="viewController">
+                <tableViewController id="1Z0-78-Skw" customClass="MissionListController" customModule="init" customModuleProvider="target" sceneMemberID="viewController">
                     <tableView key="view" clipsSubviews="YES" contentMode="scaleToFill" alwaysBounceVertical="YES" dataMode="prototypes" style="plain" separatorStyle="default" rowHeight="44" sectionHeaderHeight="28" sectionFooterHeight="28" id="UFb-CR-aiV">
                         <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
